Optionally delete identity with user profile
