+++ src/common/enums/alkemio.error.status.ts
@@ -18,6 +18,7 @@ export enum AlkemioErrorStatus {
   FORBIDDEN_POLICY = 'FORBIDDEN_POLICY',
   FORBIDDEN_LICENSE_POLICY = 'FORBIDDEN_LICENSE_POLICY',
   USER_NOT_VERIFIED = 'USER_NOT_VERIFIED',
+  USER_IDENTITY_NOT_FOUND = 'USER_IDENTITY_NOT_FOUND',
   USER_ALREADY_REGISTERED = 'USER_ALREADY_REGISTERED',
   NOT_SUPPORTED = 'NOT_SUPPORTED',
   NOT_ENABLED = 'NOT_ENABLED',
