+++ src/common/exceptions/user/user.identity.not.found.exception.ts
@@ -1,0 +1,12 @@
+import { LogContext, AlkemioErrorStatus } from '@common/enums';
+import { BaseException } from '../base.exception';
+
+export class UserIdentityNotFoundException extends BaseException {
+  constructor(
+    error: string,
+    context = LogContext.COMMUNITY,
+    code?: AlkemioErrorStatus
+  ) {
+    super(error, context, code ?? AlkemioErrorStatus.USER_IDENTITY_NOT_FOUND);
+  }
+}
