+++ src/common/utils/get.session.ts
@@ -3,9 +3,14 @@ import { FrontendApi, Session } from '@ory/kratos-client';
 import jwt_decode from 'jwt-decode';
 
 /**
- * @param kratosClient
- * @param opts
- * @throws {Error}
+ * Retrieves a session using either an authorization header or a cookie.
+ *
+ * @param kratosClient - The Kratos frontend API client.
+ * @param opts - Options for retrieving the session.
+ * @param opts.authorization - The authorization header value (optional).
+ * @param opts.cookie - The cookie value (optional).
+ * @returns A promise that resolves to a Session object or throws an error if neither authorization nor cookie is provided.
+ * @throws Will throw an error if neither authorization nor cookie is provided.
  */
 export const getSession = async (
   kratosClient: FrontendApi,
@@ -28,9 +33,12 @@ export const getSession = async (
 };
 
 /**
- * @param kratosClient
- * @param cookie
- * @throws {Error}
+ * Retrieves a session from a cookie using the Kratos client.
+ *
+ * @param kratosClient - The Kratos Frontend API client instance.
+ * @param cookie - The session cookie to retrieve the session from.
+ * @returns A promise that resolves to the session data.
+ * @throws Will throw an error if the session retrieval fails.
  */
 const getSessionFromCookie = async (
   kratosClient: FrontendApi,
@@ -46,10 +54,18 @@ const getSessionFromCookie = async (
     throw new Error(e?.message);
   }
 };
+
 /**
- * @param kratosClient
- * @param authorizationHeader
- * @throws {Error}
+ * Extracts and validates a session from the provided Authorization header.
+ *
+ * This function attempts to extract a token from the Authorization header and
+ * validate it using two different methods: `getSessionFromJwt` and
+ * `getSessionFromApiToken`. If both methods fail, an error is thrown.
+ *
+ * @param kratosClient - An instance of `FrontendApi` used to validate the token.
+ * @param authorizationHeader - The Authorization header containing the token.
+ * @returns The session object if the token is valid.
+ * @throws Will throw an error if the token is not provided or if it is invalid.
  */
 const getSessionFromAuthorizationHeader = (
   kratosClient: FrontendApi,
@@ -77,9 +93,14 @@ const getSessionFromAuthorizationHeader = (
 };
 
 /**
- * @param kratosClient
- * @param apiToken
- * @throws {Error}
+ * Retrieves a session from an API token using the Kratos client.
+ *
+ * @param kratosClient - The Kratos client instance used to fetch the session.
+ * @param apiToken - The API token used to retrieve the session.
+ * @returns A promise that resolves to a `Session` object.
+ * @throws Will throw an error if the API token is an empty string.
+ * @throws Will throw an error if the session could not be extracted from the API token.
+ * @throws Will throw an error if the session is not found for the given API token.
  */
 const getSessionFromApiToken = async (
   kratosClient: FrontendApi,
@@ -111,8 +132,14 @@ const getSessionFromApiToken = async (
 };
 
 /**
- * @param token
- * @throws {Error}
+ * Extracts a session from a JWT token.
+ *
+ * @param token - The JWT token from which to extract the session.
+ * @returns The extracted session.
+ * @throws Will throw an error if the token is empty.
+ * @throws Will throw an error if the token is a Bearer token.
+ * @throws Will throw an error if the token is not a valid JWT token.
+ * @throws Will throw an error if the Kratos session is not found in the token.
  */
 export const getSessionFromJwt = (token: string): Session | never => {
   if (!token) {
