+++ src/core/authentication/authentication.service.ts
@@ -1,11 +1,6 @@
 import { ConfigService } from '@nestjs/config';
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
-import {
-  Configuration,
-  FrontendApi,
-  IdentityApi,
-  Session,
-} from '@ory/kratos-client';
+import { Session } from '@ory/kratos-client';
 import { LogContext } from '@common/enums';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { UserService } from '@domain/community/user/user.service';
@@ -16,70 +11,44 @@ import {
 } from './ory.default.identity.schema';
 import { NotSupportedException } from '@common/exceptions';
 import { AgentService } from '@domain/agent/agent/agent.service';
-import { getBearerToken } from './get.bearer.token';
-import { SessionExtendException } from '@common/exceptions/auth';
 
 import { AgentInfoCacheService } from '@core/authentication.agent.info/agent.info.cache.service';
-import { getSession } from '@common/utils';
 import ConfigUtils from '@config/config.utils';
 import { AlkemioConfig } from '@src/types';
 import { AgentInfoMetadata } from '@core/authentication.agent.info/agent.info.metadata';
+import { KratosService } from '@services/infrastructure/kratos/kratos.service';
 
 @Injectable()
 export class AuthenticationService {
-  private readonly kratosPublicUrlServer: string;
-  private readonly adminPasswordIdentifier: string;
-  private readonly adminPassword: string;
   private readonly extendSessionMinRemainingTTL: number | undefined; // min time before session expires when it's already allowed to be extended (in milliseconds)
-  private readonly kratosIdentityClient: IdentityApi;
-  private readonly kratosFrontEndClient: FrontendApi;
 
   constructor(
     private agentCacheService: AgentInfoCacheService,
     private configService: ConfigService<AlkemioConfig, true>,
     private userService: UserService,
     private agentService: AgentService,
+    private kratosService: KratosService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER) private readonly logger: LoggerService
   ) {
-    this.kratosPublicUrlServer = this.configService.get(
-      'identity.authentication.providers.ory.kratos_public_base_url_server',
-      { infer: true }
+    const { earliest_possible_extend } = this.configService.get(
+      'identity.authentication.providers.ory',
+      {
+        infer: true,
+      }
     );
-    const {
-      kratos_public_base_url_server,
-      kratos_admin_base_url_server,
-      admin_service_account,
-      earliest_possible_extend,
-    } = this.configService.get('identity.authentication.providers.ory', {
-      infer: true,
-    });
-    this.kratosPublicUrlServer = kratos_public_base_url_server;
-    const kratosAdminUrlServer = kratos_admin_base_url_server;
-
-    this.adminPasswordIdentifier = admin_service_account.username;
-    this.adminPassword = admin_service_account.password;
 
     this.extendSessionMinRemainingTTL = this.parseEarliestPossibleExtend(
       earliest_possible_extend
-    );
-
-    this.kratosIdentityClient = new IdentityApi(
-      new Configuration({
-        basePath: kratosAdminUrlServer,
-      })
-    );
-
-    this.kratosFrontEndClient = new FrontendApi(
-      new Configuration({
-        basePath: this.kratosPublicUrlServer,
-      })
     );
   }
 
   async getAgentInfo(opts: { cookie?: string; authorization?: string }) {
     let session: Session | undefined;
     try {
-      session = await getSession(this.kratosFrontEndClient, opts);
+      session = await this.kratosService.getSession(
+        opts.authorization,
+        opts.cookie
+      );
     } catch (e) {
       return new AgentInfo();
     }
@@ -264,48 +233,25 @@ export class AuthenticationService {
   }
 
   public async extendSession(sessionToBeExtended: Session): Promise<void> {
-    const adminBearerToken = await getBearerToken(
-      this.kratosPublicUrlServer,
-      this.adminPasswordIdentifier,
-      this.adminPassword
+    const adminBearerToken = await this.kratosService.getBearerToken();
+
+    return this.kratosService.tryExtendSession(
+      sessionToBeExtended,
+      adminBearerToken
     );
-
-    return this.tryExtendSession(sessionToBeExtended, adminBearerToken);
-  }
-  // Refresh and Extend Sessions
-  // https://www.ory.sh/docs/guides/session-management/refresh-extend-sessions
-  private async tryExtendSession(
-    sessionToBeExtended: Session,
-    adminBearerToken: string
-  ): Promise<void> {
-    try {
-      /**
-       * This endpoint returns per default a 204 No Content response on success.
-       * Older Ory Network projects may return a 200 OK response with the session in the body.
-       * **Returning the session as part of the response will be deprecated in the future and should not be relied upon.**
-       * Source https://www.ory.sh/docs/reference/api#tag/identity/operation/extendSession
-       */
-      const { status } = await this.kratosIdentityClient.extendSession(
-        { id: sessionToBeExtended.id },
-        { headers: { authorization: `Bearer ${adminBearerToken}` } }
-      );
-
-      if (![200, 204].includes(status)) {
-        throw new SessionExtendException(
-          `Request to extend session ${sessionToBeExtended.id} failed with status ${status}`
-        );
-      }
-    } catch (e) {
-      if (e instanceof SessionExtendException) {
-        throw e;
-      }
-      const message = (e as Error)?.message ?? e;
-      throw new SessionExtendException(
-        `Session extend for session ${sessionToBeExtended.id} failed with: ${message}`
-      );
-    }
   }
 
+  /**
+   * Determines whether a session should be extended based on its expiration time and a minimum remaining TTL (Time To Live).
+   *
+   * @param session - The session object containing the expiration time.
+   * @returns `true` if the session should be extended, `false` otherwise.
+   *
+   * The function checks the following conditions:
+   * - If the session does not have an expiration time (`expires_at`) or the minimum remaining TTL (`extendSessionMinRemainingTTL`) is not set, it returns `false`.
+   * - If the minimum remaining TTL is set to `-1`, it returns `true`, indicating that the session can be extended at any time.
+   * - Otherwise, it calculates the session's expiry time and compares it with the current time plus the minimum remaining TTL to determine if the session should be extended.
+   */
   public shouldExtendSession(session: Session): boolean {
     if (!session.expires_at || !this.extendSessionMinRemainingTTL) {
       return false;
@@ -318,6 +264,16 @@ export class AuthenticationService {
     return Date.now() >= expiry.getTime() - this.extendSessionMinRemainingTTL;
   }
 
+  /**
+   * Parses the `earliestPossibleExtend` parameter to determine the earliest possible time to extend a session.
+   *
+   * If the `earliestPossibleExtend` is set to 'lifespan', it returns -1, allowing sessions to be refreshed during their entire lifespan.
+   * If the `earliestPossibleExtend` is a string, it attempts to parse it as a time duration in HMS format and returns the equivalent milliseconds.
+   * If the parsing fails or the input is of an unexpected type, it returns `undefined`.
+   *
+   * @param earliestPossibleExtend - The input value representing the earliest possible time to extend a session. It can be 'lifespan' or a string in HMS format.
+   * @returns The earliest possible extend time in milliseconds, -1 for 'lifespan', or `undefined` if the input is invalid.
+   */
   private parseEarliestPossibleExtend(
     earliestPossibleExtend: unknown
   ): number | undefined {
