+++ src/domain/community/user/user.service.ts
@@ -64,7 +64,7 @@ import { ContributorService } from '../contributor/contributor.service';
 import { AuthorizationPolicyType } from '@common/enums/authorization.policy.type';
 import { AccountType } from '@common/enums/account.type';
 import { AuthenticationType } from '@common/enums/authentication.type';
-import { KratosAdapter } from '@services/adapters/kratos-adapter/kratos.adapter';
+import { KratosService } from '@services/infrastructure/kratos/kratos.service';
 
 @Injectable()
 export class UserService {
@@ -81,7 +81,7 @@ export class UserService {
     private storageAggregatorService: StorageAggregatorService,
     private accountHostService: AccountHostService,
     private contributorService: ContributorService,
-    private kratosAdapter: KratosAdapter,
+    private kratosService: KratosService,
     @InjectRepository(User)
     private userRepository: Repository<User>,
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
@@ -98,9 +98,9 @@ export class UserService {
   public async getAuthenticationTypeByEmail(
     email: string
   ): Promise<AuthenticationType> {
-    const identity = await this.kratosAdapter.getIdentityByEmail(email);
+    const identity = await this.kratosService.getIdentityByEmail(email);
     if (!identity) return AuthenticationType.UNKNOWN;
-    return this.kratosAdapter.mapAuthenticationType(identity);
+    return this.kratosService.mapAuthenticationType(identity);
   }
 
   private getUserCommunicationIdCacheKey(communicationId: string): string {
@@ -423,6 +423,10 @@ export class UserService {
 
     if (user.storageAggregator) {
       await this.storageAggregatorService.delete(user.storageAggregator.id);
+    }
+
+    if (deleteData.deleteIdentity) {
+      await this.kratosService.deleteIdentityByEmail(user.email);
     }
 
     const result = await this.userRepository.remove(user as User);
