+++ src/services/infrastructure/kratos/kratos.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { KratosService } from './kratos.service';
+@Module({
+  imports: [],
+  providers: [KratosService],
+  exports: [KratosService],
+})
+export class KratosModule {}
